Wait for userData to be fetched before filling globalValues in dataLayer
Refactored the way we handle user data in the TrackingCode.twig file by introducing a new setter. This setter ensures that globalValues in the dataLayer are updated with the correct userData after it has been fetched asynchronously. This change makes sure that all user data are correctly represented, even when fetched late, improving the accuracy of tracking and data analysis.